+++ src/main/java/com/epam/estore/database/dao/interfaces/CategoryDAO.java
@@ -4,6 +4,6 @@ import com.epam.estore.entity.Category;
 
 public interface CategoryDAO {
     void insertCategory(Category category);
-    Integer getLastCategoryId();
-    void removeCategory(Integer categoryId);
+    Long getLastCategoryId();
+    void removeCategory(Long categoryId);
 }
