+++ src/main/java/com/epam/estore/database/dao/interfaces/CountryDAO.java
@@ -4,5 +4,5 @@ import com.epam.estore.entity.Country;
 import java.util.List;
 
 public interface CountryDAO {
-    List<Country> getAllCountryByLocaleId(Integer localeId);
+    List<Country> getAllCountryByLocaleId(Long localeId);
 }
