+++ src/main/java/com/epam/estore/database/dao/interfaces/ProductDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface ProductDAO {
     void insertProduct(Product product);
     Product getProductById(Long productId);
-    List<Product> getAllProductsByName(String productName, Integer localeId);
+    List<Product> getAllProductsByName(String productName, Long localeId);
     void updateCountById(Integer count, Long productId);
     void updateProduct(Product product);
     void removeProduct(Long productId);
