+++ src/main/java/com/epam/estore/database/dao/interfaces/StatusDAO.java
@@ -4,7 +4,7 @@ import com.epam.estore.entity.Status;
 import java.util.List;
 
 public interface StatusDAO {
-    Status getStatusById(Integer statusId);
-    Integer getStatusIdByNameAndLocaleId(String statusName, Integer localeId);
-    List<Status> getAllStatusByLocaleId(Integer localeId);
+    Status getStatusById(Long statusId);
+    Long getStatusIdByNameAndLocaleId(String statusName, Long localeId);
+    List<Status> getAllStatusByLocaleId(Long localeId);
 }
